更新实时数据代码修改提交

+++ tunnel-business/src/main/java/com/tunnel/business/service/dataInfo/ISdDeviceDataService.java
@@ -124,5 +124,6 @@ public interface ISdDeviceDataService {
      * @param value 数据
      * @param itemId 数据项
      */
-   void updateDeviceData(SdDevices sdDevices, String value, Integer itemId);
+   void updateDeviceData(SdDevices sdDevices, String value, Long itemId);
+
 }
